Refactor: move the additional data to types file
