+++ code-gen/src/poly_scribe_code_gen/_types.py
@@ -1,3 +1,5 @@
+from typing import Optional, TypedDict
+
 integer_types = [
     "bool",
     "char",
@@ -17,3 +19,12 @@ floating_point_types = ["float", "double", "long double"]
 std_types = ["ByteString", "string"]
 
 cpp_types = integer_types + floating_point_types + std_types
+
+
+class AdditionalData(TypedDict):
+    author_name: str
+    author_email: str
+    out_file: Optional[str]
+    year: str
+    licence: str
+    namespace: str
