+++ code-gen/src/poly_scribe_code_gen/py_gen.py
@@ -10,7 +10,7 @@ from typing import Any
 import black
 import isort
 import jinja2
-from poly_scribe_code_gen.cpp_gen import AdditionalData
+from poly_scribe_code_gen._types import AdditionalData
 
 
 def generate_python(parsed_idl: dict[str, Any], additional_data: AdditionalData, out_file: Path):
